Fix warnings in utilities

+++ src/utils/xrLC/xrSaveOGF.cpp
@@ -31,7 +31,7 @@ u16		RegisterShader		(LPCSTR T)
 	for (u32 it=0; it<pBuild->g_Shaders.size(); it++)
 		if (0==_stricmp(T,pBuild->g_Shaders[it]))	return it;
 	pBuild->g_Shaders.push_back		(xr_strdup(T));
-	return pBuild->g_Shaders.size	()-1;
+	return (unsigned short) pBuild->g_Shaders.size	()-1;
 }
 
 void	geom_batch_average	(u32 verts, u32 faces)
